Manage scenes oh so deftly
