+++ src/index.js
@@ -1,32 +1,4 @@
-import Phaser from "phaser";
-import logoImg from "./assets/logo.png";
+import Game from './game.js';
 
-const config = {
-  type: Phaser.AUTO,
-  parent: "phaser-example",
-  width: 1280,
-  height: 720,
-  scene: {
-    preload: preload,
-    create: create
-  }
-};
-
-const game = new Phaser.Game(config);
-
-function preload() {
-  this.load.image("logo", logoImg);
-}
-
-function create() {
-  const logo = this.add.image(640, 150, "logo");
-
-  this.tweens.add({
-    targets: logo,
-    y: 450,
-    duration: 2000,
-    ease: "Power2",
-    yoyo: true,
-    loop: -1
-  });
-}
+const game = new Game();
+game.scene.start('intro');
